Go and Java: update heartbeat grace period
Only apply a heartbeat grace period if maxMissedHeartbeats == 1 since
the period only applies to this case due to potential races. Also make
the grace period a function of the interval.

Remove the minimum heartbeat interval. This should not be enforced by
Frugal and should be up to the caller (e.g. Messaging SDK).