Python: don't make write a coroutine in Tornado transports
The Frugal protocol requires that all messages are framed, which means
all transports must be buffered. This means write will always just write
to a buffer and not perform IO, so it should not be a coroutine. We also
don't need to check isOpen on write since the check in flush is
sufficient.

Also clean up some exception logging.