Go: add tags to FContext
Currently, FContext headers are the equivalent of opentracing "baggage".
They get propagated across process boundaries. This adds "tags" to
FContext, which are values that do not get propagated. This will also be
useful in a number of scenarios outside of tracing, one of which is for
exposing scope topic values on the FContext.