java: Use execute if Future is unused
If an exception is thrown from a task passed to ExecutorService.submit,
the exception is captured in the returned Future.  If the Future is
unused, the exception message+stack is lost, which is problematic if
Error is thrown from a request processor.  Use ExecutorService.execute
in those cases, which allows the default exception handler to print the
exception message+stack.