java: Always set Content-Length
Commit 76890c6a added the Content-Length header for
REQUEST_ENTITY_TOO_LARGE responses (otherwise, the client hangs waiting
for the server to close the connection to signal the end of data), but
it missed updating the INTERNAL_SERVER_ERROR and BAD_REQUEST code paths.
Encountered when doing a curl on the Frugal endpoint while trying to add
an HTTP endpoint to the server.