java: Include cause in HTTP TTransportException
If an IOException occurs in FHttpTransport.makeRequest, only the message
is copied to the TTransportException, which is insufficient for
diagnosing the problem.  For example, if an invalid URL is specified,
the resulting exception is:

org.apache.thrift.transport.TTransportException: http request failed: null
	at com.workiva.frugal.transport.FHttpTransport.makeRequest(FHttpTransport.java:263)
	...

Set the cause so that the resulting exception stack trace shows:

org.apache.thrift.transport.TTransportException: http request failed: null
	at com.workiva.frugal.transport.FHttpTransport.makeRequest(FHttpTransport.java:263)
	...
Caused by: org.apache.http.client.ClientProtocolException
	at org.apache.http.impl.client.InternalHttpClient.doExecute(InternalHttpClient.java:186)
	...
Caused by: org.apache.http.ProtocolException: Target host is not specified
	...

The Apache ClientProtocolException could be improved to propagate the
underlying message, but there doesn't seem to be a good reason not to
include a cause in the TTransportException.